Update TodoList Component
