+++ client/src/Components/MainPages/Todaypage/index.tsx
@@ -1,7 +1,9 @@
 import { FC } from 'react';
 import styled from 'styled-components';
 import { useSelector } from 'react-redux';
-import { StyledBackGroundColorContainer, RootState } from '../../../interfaces';
+import { StyledBackGroundColorContainer, RootState, SingleTodoInterface } from '../../../interfaces';
+import TodoList from '../Components/TodoList';
+import moment from 'moment';
 
 const TodayPageContainer = styled.div<StyledBackGroundColorContainer>`
     background-color: ${({ backgroundColor }) => backgroundColor};
@@ -9,11 +11,28 @@ const TodayPageContainer = styled.div<StyledBackGroundColorContainer>`
     width: 100%;
 `;
 
+const TodoContainer = styled.main`
+    width: 100%;
+    height: 100%;
+    padding: 10px;
+    color: white;
+`;
+
 const TodayPage: FC = () => {
     const { backgroundColor } = useSelector((state: RootState) => state.AppColor);
     const { total: totalTodo } = useSelector((state: RootState) => state.Todo);
 
-    return <TodayPageContainer backgroundColor={backgroundColor}></TodayPageContainer>;
+    console.log(totalTodo);
+
+    const todayTodoList = totalTodo.filter((todo: SingleTodoInterface) => todo.time === moment().format('YYYYMMDD'));
+
+    return (
+        <TodayPageContainer backgroundColor={backgroundColor}>
+            <TodoContainer>
+                <TodoList title={'오늘 할일'} totalTodo={todayTodoList} />
+            </TodoContainer>
+        </TodayPageContainer>
+    );
 };
 
 export default TodayPage;
